pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "nottodbox"
version = "0.1.0"
authors = [{name="MuKonqi (Muhammed S.)"}]
description = "Organize notes, to-dos and diaries"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Intended Audience :: End Users/Desktop",
  "Topic :: Text Editors",
  "Topic :: Text Processing",
  "Topic :: Utilities",
  "Programming Language :: Python",
  "Operating System :: OS Independent",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"
]
dependencies = ["PySide6-Essentials"]

[project.urls]
Repository = "https://github.com/mukonqi/nottodbox"
Issues = "https://github.com/mukonqi/nottodbox/issues"

[project.scripts]
nottodbox = "nottodbox.scripts.application:main"

[tool.setuptools]
packages = ["nottodbox", "nottodbox.scripts", "nottodbox.scripts.resources", "nottodbox.scripts.widgets"]

[tool.setuptools.package-data]
nottodbox = ["LICENSE.txt", "color-schemes/*"]